copernicoaches/jano/pyproject.toml
[project]
name = "jano"
version = "0.1.0"
description = "Bidirectional Word ↔ Markdown converter"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.1",
    "markdown-it-py>=3.0",
    "fastmcp>=2.0",
]

[project.scripts]
jano = "jano.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
dev-mode-dirs = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/jano"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
